\name{print.pairwise.htest}
\alias{print.pairwise.htest}
\title{Print method for pairwise tests}
\description{
Display results of pairwise comparison procedures
}
\usage{
\method{print}{pairwise.htest}(x, \dots)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{ Object of class \code{"pairwise.htest"} }
  \item{\dots}{further arguments to be passed to or from methods.}  
}

\value{
  None
}
\seealso{ \code{\link{pairwise.t.test}}, et al.}

\keyword{ htest }%-- one or more ...
